Brings the quality pass, Movies pagination, the Featured reel and kids
profiles (0.43.0) onto main. The branch already carried its own version of
c2e7578's refresh-before-resume fix (refreshState, the same five tests), so
every conflict resolves to the branch side.

+++ crates/mediagram-core/src/shows/mod.rs
@@ -44,7 +44,7 @@ pub fn upsert(conn: &Connection, row: &TitleDetailsRow) -> rusqlite::Result<()> 
         params![
             SOURCE,
             kind_key(row.kind),
-            row.id as i64,
+            row.id,
             row.lang,
             row.overview,
             row.tagline,
@@ -72,7 +72,7 @@ pub fn get(conn: &Connection, kind: Kind, id: u64) -> rusqlite::Result<Option<Ti
                     total_seasons, total_episodes, {certification}
                FROM shows WHERE source = ?1 AND kind = ?2 AND id = ?3"
         ),
-        params![SOURCE, kind_key(kind), id as i64],
+        params![SOURCE, kind_key(kind), id],
         |row| {
             Ok(TitleDetailsRow {
                 kind,
